refactor: update the validators' extension logics to reduce redundancy
